+++ CHANGELOG.md
@@ -2,6 +2,9 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bump runtimeVersion to `2026.02.10`.
+ * Upgraded stable Dart analysis SDK to `3.11.0`
+ * Upgraded stable Flutter analysis SDK to `3.38.9`.
 
  * Upgraded _Container Optimized OS_ to [Milestone 125](https://docs.cloud.google.com/container-optimized-os/docs/release-notes/m125)